Add support for square keys
Add support for keys that are as wide as the height of the current key
row high (i.e. keys that have the shape of a square) or a multiple of
the height. For such keys calculate the key width dynamically instead of
setting a relative width in the layout header file.

To use this feature, you need to set the member shape of a key in a
custom layout to OneSquare, TwoSquare, or ThreeSquare. If you don't
explicitly set the shape, the default is still DynamicWidth.