anihortes: Add Home and End keys
The Home and End keys are desperately needed on the primary layout. Put
those in the most intuitive position which, for me, is swiping left or
right on the outer keys of the middle row (h and r).

The parenthesis keys are currently occupying those positions, so move
them to swiping north-west on h and north-east on r. The curly braces
are currently occupying these positions, but they remain available via
drag-return on the brackets (south-west on h and south-east on r), so
the trade-off seems acceptable to me.